Adjusted folder format and files
