+++ src/main/java/com/sbi/TestJPA/models/Analisis.java
@@ -1,28 +1,25 @@
-package com.sbi.TestJPA.model;
+package com.sbi.TestJPA.models;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Column;
-import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import javax.persistence.JoinColumn;
 
 @Entity
-public class Noticia {
+public class Analisis {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Integer id;
 	
-	@Lob
-	@Column
-	private String noticia;
-	
 	@ManyToOne
 	@JoinColumn(name="idpeticion", referencedColumnName="id")
 	private Peticion idpeticion;
+	
+	private String palabra;
+	private Integer contador;
 	
 	
 	public Integer getId() {
@@ -31,23 +28,35 @@ public class Noticia {
 	public void setId(Integer id) {
 		this.id = id;
 	}
-	public String getNoticia() {
-		return noticia;
-	}
-	public void setNoticia(String noticia) {
-		this.noticia = noticia;
-	}
 	public Peticion getIdpeticion() {
 		return idpeticion;
 	}
 	public void setIdpeticion(Peticion idpeticion) {
 		this.idpeticion = idpeticion;
 	}
-	
-	public Noticia(Peticion peti, String strNoti) {
-		idpeticion = peti;
-		noticia = strNoti;
+	public String getPalabra() {
+		return palabra;
+	}
+	public void setPalabra(String palabra) {
+		this.palabra = palabra;
+	}
+	public Integer getContador() {
+		return contador;
+	}
+	public void setContador(Integer contador) {
+		this.contador = contador;
 	}
 	
+	public Analisis(Peticion peti, String strPalabra) {
+		idpeticion = peti;
+		palabra = strPalabra;
+		contador = 0;
+	}
+	
+	public Analisis() {}
+	
+	public void incrementar() {
+		contador++;
+	}
 
 }
