+++ src/main/java/com/sbi/TestJPA/models/LLamadoWrap.java
@@ -1,4 +1,4 @@
-package com.sbi.TestJPA.model;
+package com.sbi.TestJPA.models;
 
 import java.util.List;
