+++ src/main/java/com/sbi/TestJPA/models/Llamado.java
@@ -1,4 +1,4 @@
-package com.sbi.TestJPA.model;
+package com.sbi.TestJPA.models;
 
 public class Llamado {
