+++ src/main/java/com/sbi/TestJPA/models/Noticia.java
@@ -1,25 +1,28 @@
-package com.sbi.TestJPA.model;
+package com.sbi.TestJPA.models;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.Column;
+import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import javax.persistence.JoinColumn;
 
 @Entity
-public class Analisis {
+public class Noticia {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Integer id;
 	
+	@Lob
+	@Column
+	private String noticia;
+	
 	@ManyToOne
 	@JoinColumn(name="idpeticion", referencedColumnName="id")
 	private Peticion idpeticion;
-	
-	private String palabra;
-	private Integer contador;
 	
 	
 	public Integer getId() {
@@ -28,35 +31,21 @@ public class Analisis {
 	public void setId(Integer id) {
 		this.id = id;
 	}
+	public String getNoticia() {
+		return noticia;
+	}
+	public void setNoticia(String noticia) {
+		this.noticia = noticia;
+	}
 	public Peticion getIdpeticion() {
 		return idpeticion;
 	}
 	public void setIdpeticion(Peticion idpeticion) {
 		this.idpeticion = idpeticion;
 	}
-	public String getPalabra() {
-		return palabra;
-	}
-	public void setPalabra(String palabra) {
-		this.palabra = palabra;
-	}
-	public Integer getContador() {
-		return contador;
-	}
-	public void setContador(Integer contador) {
-		this.contador = contador;
-	}
 	
-	public Analisis(Peticion peti, String strPalabra) {
+	public Noticia(Peticion peti, String strNoti) {
 		idpeticion = peti;
-		palabra = strPalabra;
-		contador = 0;
+		noticia = strNoti;
 	}
-	
-	public Analisis() {}
-	
-	public void incrementar() {
-		contador++;
-	}
-
 }
