+++ src/main/java/com/sbi/TestJPA/repositories/AnalisisRepo.java
@@ -1,11 +1,11 @@
-package com.sbi.TestJPA.repository;
+package com.sbi.TestJPA.repositories;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import com.sbi.TestJPA.model.Analisis;
+import com.sbi.TestJPA.models.Analisis;
 
 @Repository
 public interface AnalisisRepo  extends JpaRepository<Analisis, Integer>{
