+++ src/main/java/com/sbi/TestJPA/repositories/NoticiaRepo.java
@@ -1,8 +1,8 @@
-package com.sbi.TestJPA.repository;
+package com.sbi.TestJPA.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.sbi.TestJPA.model.Noticia;
+import com.sbi.TestJPA.models.Noticia;
 
 
 @Repository
