+++ src/main/java/com/sbi/TestJPA/utils/Constantes.java
@@ -1,4 +1,4 @@
-package com.sbi.TestJPA.util;
+package com.sbi.TestJPA.utils;
 
 
 public class Constantes {
