+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@
 spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
-spring.datasource.password=
+spring.datasource.password=sa
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 # Hbilitar la consola H2
 spring.h2.console.enabled=true
